+++ packages/admin/admin-generator/src/commands/generate/generateForm/generateFormField.ts
@@ -4,6 +4,7 @@ import { type FormConfig, type FormFieldConfig, type GQLDocumentConfigMap } from
 import { camelCaseToHumanReadable } from "../utils/camelCaseToHumanReadable";
 import { convertConfigImport } from "../utils/convertConfigImport";
 import { type Imports } from "../utils/generateImportsCode";
+import { generateFormattedMessage } from "../utils/intl";
 import { isFieldOptional } from "../utils/isFieldOptional";
 import { isGeneratorConfigCode, isGeneratorConfigImport } from "../utils/runtimeTypeGuards";
 import { generateAsyncSelect } from "./asyncSelect/generateAsyncSelect";
@@ -97,15 +98,20 @@ export function generateFormField({
             ${config.endAdornment ? `endAdornment={<InputAdornment position="end">${endAdornment.adornmentString}</InputAdornment>}` : ""}
             ${
                 config.helperText
-                    ? `helperText={<FormattedMessage id=` +
-                      `"${formattedMessageRootId}.${name}.helperText" ` +
-                      `defaultMessage="${config.helperText}" />}`
+                    ? `helperText={${generateFormattedMessage({
+                          config: config.helperText,
+                          id: `${formattedMessageRootId}.${name}.helperText`,
+                          type: "jsx",
+                      })}}`
                     : ""
             }
             ${validateCode}
         />`;
         if (!required && !config.readOnly) {
             formValueConfig.formValueToGqlInputCode = `$fieldName ?? null`;
+        }
+        if (config.initialValue !== undefined) {
+            formValueConfig.defaultInitializationCode = JSON.stringify(config.initialValue);
         }
     } else if (config.type == "number") {
         code = `
@@ -145,6 +151,9 @@ export function generateFormField({
             type: "string",
         };
         formValueConfig.initializationCode = `${initializationAssignment}`;
+        if (config.initialValue !== undefined) {
+            formValueConfig.defaultInitializationCode = JSON.stringify(config.initialValue);
+        }
     } else if (config.type === "numberRange") {
         code = `
             <Field
@@ -171,6 +180,9 @@ export function generateFormField({
             />`;
 
         formFragmentFields = [`${name}.min`, `${name}.max`];
+        if (config.initialValue !== undefined) {
+            formValueConfig.defaultInitializationCode = JSON.stringify(config.initialValue);
+        }
     } else if (config.type == "boolean") {
         code = `<CheckboxField
                         label={${fieldLabel}}
@@ -187,7 +199,7 @@ export function generateFormField({
                         }
                         ${validateCode}
                     />`;
-        formValueConfig.defaultInitializationCode = `false`;
+        formValueConfig.defaultInitializationCode = config.initialValue ? "true" : "false";
     } else if (config.type == "date") {
         imports.push({
             name: "Future_DatePickerField",
@@ -214,6 +226,9 @@ export function generateFormField({
             />`;
         if (!required && !config.readOnly) {
             formValueConfig.formValueToGqlInputCode = `$fieldName ?? null`;
+        }
+        if (config.initialValue !== undefined) {
+            formValueConfig.defaultInitializationCode = JSON.stringify(config.initialValue);
         }
     } else if (config.type == "dateTime") {
         imports.push({
@@ -246,6 +261,9 @@ export function generateFormField({
         };
         if (!config.readOnly) {
             formValueConfig.formValueToGqlInputCode = required ? `$fieldName.toISOString()` : `$fieldName ? $fieldName.toISOString() : null`;
+        }
+        if (config.initialValue !== undefined) {
+            formValueConfig.defaultInitializationCode = `new Date("${config.initialValue.toISOString()}")`;
         }
     } else if (config.type == "block") {
         code = `<Field name="${nameWithPrefix}" isEqual={isEqual} label={${fieldLabel}} variant="horizontal" fullWidth>
@@ -369,6 +387,9 @@ export function generateFormField({
                                     }`;
                                 })}]}
                             />`;
+        }
+        if (config.initialValue !== undefined) {
+            formValueConfig.defaultInitializationCode = JSON.stringify(config.initialValue);
         }
     } else {
         throw new Error(`Unsupported type`);