+++ packages/admin/admin-generator/src/commands/generate/generateGrid/__tests__/generateGrid.test.ts
@@ -173,4 +173,86 @@ describe("generateGrid", () => {
 
         expect(result.code).toMatchSnapshot();
     });
+
+    it("should generate empty headerName without formattedMessage", () => {
+        const config: GridConfig<Book> = {
+            type: "grid",
+            gqlType: "Book",
+            query: "books",
+            excelExport: true,
+            columns: [
+                {
+                    type: "text",
+                    name: "title",
+                    headerName: "",
+                },
+            ],
+        };
+
+        const result = generateGrid(
+            {
+                exportName: "BooksGrid",
+                baseOutputFilename: "BooksGrid",
+                targetDirectory: "/test",
+                gqlIntrospection: introspection,
+            },
+            config,
+        );
+
+        expect(result.code).toMatch(/field: "title",\s*headerName: "",/);
+    });
+    it("should generate a grid with density setting", () => {
+        const config: GridConfig<Book> = {
+            type: "grid",
+            gqlType: "Book",
+            density: "compact",
+            columns: [
+                {
+                    type: "text",
+                    name: "title",
+                },
+            ],
+        };
+
+        const result = generateGrid(
+            {
+                exportName: "BooksGrid",
+                baseOutputFilename: "BooksGrid",
+                targetDirectory: "/test",
+                gqlIntrospection: introspection,
+            },
+            config,
+        );
+
+        expect(result.code).toContain('density="compact"');
+    });
+
+    it("should generate custom text for delete action in crudContextMenu", () => {
+        const config: GridConfig<Book> = {
+            type: "grid",
+            gqlType: "Book",
+            query: "books",
+            columns: [
+                {
+                    type: "text",
+                    name: "title",
+                },
+            ],
+            crudContextMenu: {
+                deleteText: "Extinguish",
+            },
+        };
+
+        const result = generateGrid(
+            {
+                exportName: "BooksGrid",
+                baseOutputFilename: "BooksGrid",
+                targetDirectory: "/test",
+                gqlIntrospection: introspection,
+            },
+            config,
+        );
+
+        expect(result.code).toMatchSnapshot();
+    });
 });