+++ packages/admin/admin/src/dataGrid/muiGridSortToGql.tsx
@@ -8,7 +8,7 @@ type SortEntry = {
 };
 
 export function muiGridSortToGql(sortModel?: GridSortModel, columns?: GridColDef[]) {
-    if (!sortModel) return undefined;
+    if (!sortModel || sortModel.length === 0) return undefined;
 
     const sortFieldMapping: Record<string, string[]> = {};