+++ packages/admin/cms-admin/src/pages/pageTree/PageTree.tsx
@@ -398,7 +398,7 @@ function getLinkedPages(pages: PageTreePage[], index: number): [PageTreePage | u
 
 const PageTreeWithRef = forwardRef(PageTree);
 
-export { PageTreeWithRef as PageTree, PageTreeRefApi };
+export { PageTreeWithRef as PageTree, type PageTreeRefApi };
 
 const Root = styled("div")`
     height: 100%;